+++ Frontend/HLC/src/views/FavoriteMovieView.vue
@@ -1,12 +1,13 @@
 <template>
-  <div>
-    <h1>내가 좋아하는 영화들</h1>
-    <FavCarousel />
+  <div class="p-5">
+    <h3>내가 찜한 영화 목록</h3>
+    <hr style="border: solid 2px #fff;">
+    <FavMovies />
   </div>
 </template>
 
 <script setup>
-import FavCarousel from '@/components/movie/FavCarousel.vue'
+import FavMovies from '@/components/movie/FavMovies.vue'
 </script>
 
 <style scoped>
